+++ package.json
@@ -27,6 +27,7 @@
     "react-native-gesture-handler": "2.8.0",
     "react-native-lottie-splash-screen": "1.0.1",
     "react-native-pager-view": "6.1.1",
+    "react-native-phone-number-input-formatted": "^2.2.8",
     "react-native-reanimated": "2.13.0",
     "react-native-safe-area-context": "4.4.1",
     "react-native-screens": "3.18.1",
