+++ src/components/DateInputFieldWithIcon/DateInputFieldWithIcon.styles.js
