--- src/components/DateInputIcon/index.js
@@ -1,1 +1,0 @@
-export {default} from './DateInputIcon';
