--- src/components/PhoneInputIcon/index.js
@@ -1,1 +1,0 @@
-export {default} from './PhoneInputIcon';
