+++ src/components/TextInputFieldWithIcon/TextInputFieldWithIcon.styles.js
